requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
greenlet>=3.0.0
pydantic==2.10.6
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.10
requests==2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dateutil==2.8.2
asyncpg==0.29.0
python-dotenv==1.0.0
aiohttp>=3.11.11,<4.0.0
pydantic-settings==2.7.1
werkzeug==3.1.3

# Redis for caching and message broker
redis==4.6.0
hiredis==2.3.2

# Performance and monitoring
psutil==5.9.8

# AI/ML Dependencies
textblob==0.19.0
nltk==3.9.1
sentence-transformers==3.4.1
prophet==1.1.6
scikit-learn>=1.3.0

# LLM Provider SDKs
openai>=1.0.0
anthropic>=0.25.0
groq>=0.4.0
cohere>=5.0.0
tenacity>=8.2.0

# Google API Dependencies
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-analytics-data==0.18.18

# Web Scraping & HTML Parsing
beautifulsoup4==4.13.4
soupsieve==2.7
tldextract==5.3.0
playwright>=1.40.0
bleach==6.1.0
dnspython==2.6.1

# PDF Export Dependencies
reportlab>=4.0.4
matplotlib>=3.7.1
seaborn>=0.12.2
pillow>=9.0.0
asyncio>=3.4.3

# Test Dependencies
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.1.1
pytest-mock==3.14.0
aiosqlite==0.21.0
httpx==0.28.1
backoff==2.2.1
ratelimit==2.2.1

# Celery for background tasks
celery[redis]==5.3.4
flower==2.0.1

# API Clients
google-search-results==2.4.2
serpapi==0.1.5

# WebSocket support
websockets==12.0

# MinIO Object Storage
minio==7.2.5

# GraphDB / Neo4j
neo4j==5.15.0

# Additional dependencies for new features
croniter==2.0.1  # For cron expression parsing (scheduled reports)
langdetect==1.0.9  # For language detection (i18n)
